Move from project to device.
